move compendium to own file, rename lookupFuncs
