+++ cogs5e/initTracker.py
@@ -10,7 +10,7 @@ from discord.ext import commands
 from discord.ext.commands import NoPrivateMessage
 
 from cogs5e.funcs import attackutils, checkutils, targetutils
-from gamedata.lookupFuncs import select_monster_full, select_spell_full
+from gamedata.lookuputils import select_monster_full, select_spell_full
 from cogs5e.funcs.scripting import helpers
 from cogs5e.models.character import Character
 from cogs5e.models.embeds import EmbedWithAuthor, EmbedWithCharacter
