+++ gamedata/compendium.py
@@ -1,10 +1,4 @@
-"""
-Created on Jan 13, 2017
-
-@author: andrew
-"""
 import asyncio
-import itertools
 import json
 import logging
 import os
@@ -12,18 +6,11 @@ import os
 import newrelic.agent
 
 from cogs5e.models.background import Background
-from cogs5e.models.errors import NoActiveBrew
-from cogs5e.models.homebrew import Tome
-from cogs5e.models.homebrew.bestiary import Bestiary
 from cogs5e.models.monster import Monster
 from cogs5e.models.race import Race
 from cogs5e.models.spell import Spell
-from cogsmisc.stats import Stats
 from utils import config
-from utils.functions import parse_data_entry, search_and_select
-
-HOMEBREW_EMOJI = "<:homebrew:434140566834511872>"
-HOMEBREW_ICON = "https://avrae.io/assets/img/homebrew.png"
+from utils.functions import parse_data_entry
 
 log = logging.getLogger(__name__)
 
@@ -44,7 +31,6 @@ class Compendium:
         self.spells = []
         self.rule_references = []
         self.srd_backgrounds = []
-        self.srd_items = []
         self.srd_races = []
         self.srd_spells = []
         self.subclasses = []
@@ -97,8 +83,9 @@ class Compendium:
         self.names = self.read_json('names.json', [])
         self.rule_references = self.read_json('srd-references.json', [])
         self.srd_backgrounds = self.read_json('srd-backgrounds.json', [])
-        self.srd_items = self.read_json('srd-items.json', [])
+        self.items = self.read_json('srd-items.json', [])
         self.srd_races = self.read_json('srd-races.json', [])
+        self.rfeats = self.read_json('srd-racefeats', [])
         self.srd_spells = self.read_json('srd-spells.json', [])
 
         # Dictionary!
@@ -115,8 +102,9 @@ class Compendium:
         self.names = lookup.get('names', [])
         self.rule_references = lookup.get('srd-references', [])
         self.srd_backgrounds = lookup.get('srd-backgrounds', [])
-        self.srd_items = lookup.get('srd-items', [])
+        self.items = lookup.get('srd-items', [])
         self.srd_races = lookup.get('srd-races', [])
+        self.rfeats = lookup.get('srd-racefeats', [])
         self.srd_spells = lookup.get('srd-spells', [])
 
         # Dictionary!
@@ -129,21 +117,8 @@ class Compendium:
         self.monster_mash = [Monster.from_data(m) for m in self.monsters]
         self.spells = [Spell.from_data(s) for s in self.srd_spells]
 
-        self.items = [i for i in self.srd_items if i.get('type') != '$']
-
-        self.rfeats = self._load_rfeats()
         self.subclasses = self._load_subclasses()
         self._load_nsrd_names()
-
-    def _load_rfeats(self):
-        ret = []
-        for race in self.srd_races:
-            for entry in race['entries']:
-                if isinstance(entry, dict) and 'name' in entry:
-                    temp = {'name': "{}: {}".format(race['name'], entry['name']),
-                            'text': parse_data_entry(entry['entries']), 'srd': race['srd']}
-                    ret.append(temp)
-        return ret
 
     def _load_subclasses(self):
         s = []
@@ -176,85 +151,6 @@ class Compendium:
             log.warning("File not found: {}".format(filepath))
         log.debug("Loaded {} things from file {}".format(len(data), filename))
         return data
-
-
-compendium = Compendium()
-
-
-# ---- helper ----
-def get_homebrew_formatted_name(named):
-    if named.source == 'homebrew':
-        return f"{named.name} ({HOMEBREW_EMOJI})"
-    return named.name
-
-
-# ----- Monster stuff
-async def select_monster_full(ctx, name, cutoff=5, return_key=False, pm=False, message=None, list_filter=None,
-                              return_metadata=False, extra_choices=None, selectkey=None):
-    """
-    Gets a Monster from the compendium and active bestiary/ies.
-    """
-    try:
-        bestiary = await Bestiary.from_ctx(ctx)
-        await bestiary.load_monsters(ctx)
-        custom_monsters = bestiary.monsters
-        bestiary_id = bestiary.id
-    except NoActiveBrew:
-        custom_monsters = []
-        bestiary_id = None
-    choices = list(itertools.chain(compendium.monster_mash, custom_monsters))
-    if ctx.guild:
-        async for servbestiary in Bestiary.server_bestiaries(ctx):
-            if servbestiary.id == bestiary_id:
-                continue
-            await servbestiary.load_monsters(ctx)
-            choices.extend(servbestiary.monsters)
-
-    await Stats.increase_stat(ctx, "monsters_looked_up_life")
-
-    # #881
-    if extra_choices:
-        choices.extend(extra_choices)
-    if selectkey is None:
-        selectkey = get_homebrew_formatted_name
-
-    return await search_and_select(ctx, choices, name, lambda e: e.name, cutoff, return_key, pm, message, list_filter,
-                                   selectkey=selectkey, return_metadata=return_metadata)
-
-
-# ---- SPELL STUFF ----
-async def select_spell_full(ctx, name, *args, extra_choices=None, **kwargs):
-    """
-    Gets a Spell from the compendium and active tome(s).
-
-    :rtype: :class:`~cogs5e.models.spell.Spell`
-    """
-    choices = await get_spell_choices(ctx)
-    await Stats.increase_stat(ctx, "spells_looked_up_life")
-
-    # #881
-    if extra_choices:
-        choices.extend(extra_choices)
-    if 'selectkey' not in kwargs:
-        kwargs['selectkey'] = get_homebrew_formatted_name
-
-    return await search_and_select(ctx, choices, name, lambda e: e.name, *args, **kwargs)
-
-
-async def get_spell_choices(ctx):
-    try:
-        tome = await Tome.from_ctx(ctx)
-        custom_spells = tome.spells
-        tome_id = tome.id
-    except NoActiveBrew:
-        custom_spells = []
-        tome_id = None
-    choices = list(itertools.chain(compendium.spells, custom_spells))
-    if ctx.guild:
-        async for servtome in Tome.server_active(ctx):
-            if servtome.id != tome_id:
-                choices.extend(servtome.spells)
-    return choices
 
 
 # ==== nsrd helper class ====
