+++ tests/static/compendium/srd-racefeats.json
@@ -1,0 +1,22 @@
+[
+  {
+    "name": "Human: Age",
+    "text": "Humans reach adulthood in their late teens and live less than a century.",
+    "srd": true
+  },
+  {
+    "name": "Human: Alignment",
+    "text": "Humans tend toward no particular alignment. The best and the worst are found among them.",
+    "srd": true
+  },
+  {
+    "name": "Human: Size",
+    "text": "Humans vary widely in height and build, from barely 5 feet to well over 6 feet tall. Regardless of your position in that range, your size is Medium.",
+    "srd": true
+  },
+  {
+    "name": "Human: Languages",
+    "text": "You can speak, read, and write Common and one extra language of your choice. Humans typically learn the languages of other peoples they deal with, including obscure dialects. They are fond of sprinkling their speech with words borrowed from other tongues: Orc curses, Elvish musical expressions, Dwarvish military phrases, and so on.",
+    "srd": true
+  }
+]
